% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data.r
\name{input_data}
\alias{input_data}
\title{Evaluate and Visualize Hydrometric Data Availability}
\usage{
input_data(stream_stages, rating_curves, obs_streamflows)
}
\arguments{
\item{stream_stages}{a data frame with three columns: station name
(character), time istant of observation (character), stream stage
observation (numeric)}

\item{rating_curves}{a data frame with six columns: station name
(character), start date of applicability (character), end date of
applicability (character), minimum applicable stage (numeric), maximum
applicable stage (numeric), rating curve equation (character)}

\item{obs_streamflows}{a data frame with three columns: station name
(stations in common with stream_stages must match), time istant of
observation (character), streamflow observation (numeric)}
}
\value{
store output in the internal environment \code{.internal_env}:
\itemize{
\item \code{.internal_env$input_data}: List of data frames: \code{stream_stages},\code{rating_curves}, \code{obs_streamflows}
\item \code{.internal_env$DA_matrices}: List of matrices: \code{stream_stages} and \code{obs_streamflow} availability matrix,
\code{rating_curves} binary matrix
}
}
\description{
Computes stream stages, rating curves and streamflows availability matrices
and plot them as heatmaps
}
\details{
Parse a date-time or date string in common international format,
match rating curve to stream stages by extracting \code{location} from
\code{station@location}, calculates annual data availability for stream stages and
streamflows as proportions of expected values based on timestamp (5 min, 30
min, 1 hour), constructs three availability matrices (stream stages, rating
curves, streamflows) by stations and years and plots stream stages and daily
streamflow availability in grayscale color ("white": no data, "black":
complete year)
}
\note{
This function uses \code{lubridate} and \code{plot.matrix} assumes a global
environment variable \code{.internal_env} exists
}
\examples{
# Assuming you have stream_stages, rating_curves and obs_streamflows dataframes loaded
input_data(stream_stages, rating_curves, obs_streamflows)

# Access availability matrices
.internal_env$DA_matrices$stream_stages
.internal_env$DA_matrices$rating_curves
.internal_env$DA_matrices$obs_streamflows

}
