% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_internal_data.R
\name{use_internal_data}
\alias{use_internal_data}
\title{Extract lists in the internal environment}
\usage{
use_internal_data(list_name, dataset_name)
}
\arguments{
\item{list_name}{character string of name list}

\item{dataset_name}{character string of name data frame}
}
\value{
return the requested dataframe
if no matching is found (list or dataframe), the function stops with an error
}
\description{
Extract lists in the internal environment
}
\examples{
\dontrun{
# Assuming you have stored stream_stages dataframe inside DA_matrices list
# Access availability matrices
use_internal_data(list_name = "DA_matrices",dataset_name = "stream_stages")
}

}
